requirements.txt
pandas>=1.3.0
requests>=2.25.0
python-dotenv>=0.19.0
urllib3>=1.26.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.12.0